Various updates for 2.7.0-RC3.
Fixed problems with the build on some systems.

Applications now use native threads instead of OpenMP threads.

Allow for spurious wake-ups in thread barrier.

Allow DFT imager mode to be used without a GPU.

Added option to change current work dir from GUI.

Allowed FITS images to be read if they have a diagonal CD matrix.

Updated install documentation.

Added packaging scripts.

Added third-party licenses to documentation.

+++ apps/oskar_convert_ecef_to_enu_main.cpp
@@ -68,7 +68,7 @@ int main(int argc, char** argv)
             num_points, &status);
     oskar_Mem *enu_z = oskar_mem_create(OSKAR_DOUBLE, OSKAR_CPU,
             num_points, &status);
-    oskar_convert_ecef_to_enu(num_points,
+    oskar_convert_ecef_to_enu((int)num_points,
             oskar_mem_double_const(ecef_x, &status),
             oskar_mem_double_const(ecef_y, &status),
             oskar_mem_double_const(ecef_z, &status),
