+++ apps/oskar_vis_upgrade_format_main.cpp
@@ -57,7 +57,7 @@ int main(int argc, char** argv)
     string out_path;
     vector<string> in_files = opt.get_input_files(1);
     bool verbose = opt.is_set("-q") ? false : true;
-    int num_in_files = in_files.size();
+    int num_in_files = (int)in_files.size();
 
     // Print if verbose.
     if (verbose)
