+++ oskar/mem/src/oskar_mem_add.c
@@ -100,7 +100,7 @@ void oskar_mem_add(oskar_Mem* out, const oskar_Mem* in1, const oskar_Mem* in2,
         else if (location == OSKAR_GPU)
         {
 #ifdef OSKAR_HAVE_CUDA
-            oskar_mem_add_cuda_d(num_elements, bb, cc, aa);
+            oskar_mem_add_cuda_d((int) num_elements, bb, cc, aa);
             oskar_device_check_error(status);
 #else
             *status = OSKAR_ERR_CUDA_NOT_AVAILABLE;
@@ -133,7 +133,7 @@ void oskar_mem_add(oskar_Mem* out, const oskar_Mem* in1, const oskar_Mem* in2,
         else if (location == OSKAR_GPU)
         {
 #ifdef OSKAR_HAVE_CUDA
-            oskar_mem_add_cuda_f(num_elements, bb, cc, aa);
+            oskar_mem_add_cuda_f((int) num_elements, bb, cc, aa);
             oskar_device_check_error(status);
 #else
             *status = OSKAR_ERR_CUDA_NOT_AVAILABLE;
